scheduler: reap task slots + exit-status ring + WaitTask UAF fix (ADR-034 Phase B.2-B.4)
Closes the task-slot leak and the SYS_WAIT_TASK use-after-free, building on
the generation-carrying TaskId from Phase B.1. The per-CPU reaper now also
frees the dead task's scheduler slot; a parent collects the exit code from a
generation-keyed ring that outlives the slot. The spawn/wait ABI widens to a
full (slot, generation) handle so a reaped+reused slot is detected, never
silently collected as a different task.

What:
  - B.2 reaper slot reclaim (src/reaper.rs, src/scheduler/mod.rs,
    src/syscalls/dispatcher.rs): ReclaimItem carries the dead TaskId;
    drain_local frees the slot via Scheduler::reap_slot (frees iff the slot
    still holds that exact (slot, generation) Terminated task) and bumps the
    global generation, under the local SCHEDULER lock (free + bump atomic, so
    no create on this CPU grabs the slot with the stale generation). The slot
    is local to the reaper — the dying task enqueued to its own CPU and a
    Terminated+purged task cannot migrate.
  - B.3 exit ring + WaitTask rewrite (src/scheduler/task.rs, src/lib.rs,
    src/syscalls/dispatcher.rs): a global TaskExitRing latches
    (slot, generation, exit_code, parent) during exit, before the reaper frees
    the slot. handle_wait_task: ring fast-path -> cross-CPU authorize+state
    (task_parent_and_state) -> H5 cli-before-block with a lost-wakeup guard
    (re-check the ring under the scheduler lock the waker must take) -> yield
    -> ring read. Selective wake: BlockReason::ChildWait(child_slot) +
    wake_child_waiter wakes only the parent waiting on the exiting child, not
    a parent waiting on a sibling (the spurious-wake the unconditional wake
    allowed). handle_exit records the ring before the selective wake.
  - B.4 ABI widen (src/syscalls/dispatcher.rs, user/libsys, user/shell,
    user/terminal-window): spawn returns new_task_id.as_raw(); wait_task takes
    the full u64 handle via from_raw. The shell and terminal-window were
    silently truncating the handle to the slot (dropping the generation) —
    now carry the full handle.
  - 8 host unit tests (reap_slot incl. stale-generation + idempotent;
    wake_child_waiter matching/sibling/unblocked; TaskExitRing record/lookup +
    generation-mismatch + age-out). SCHEDULER.md synced (primitives, ChildWait,
    test count 35 -> 43). ASSUMPTIONS.md: TASK_EXIT_RING_SIZE + TaskId
    generation-width rows.

Adversarial verification: a 7-reviewer workflow (lost-wakeup, slot-UAF,
lock-order, generation, ABI, reaper-double-free, cross-CPU-bounds) found and
this commit fixes:
  - HIGH: WaitTask returned InvalidArg (garbage exit code) when the child
    exited+reaped in the gap between the fast-path ring lookup and the resolve
    (the shell-waits-on-a-game common case). Fixed: the resolve's not-found
    arm re-checks the ring before erroring.
  - LOW: task_parent_and_state could spuriously not-find a live child the load
    balancer migrates mid-resolve. Fixed: a bounded (<=2) retry on a re-read CPU.
  - LOW: generation-bit-31 / i64-spawn-handle horizon — documented (ASSUMPTIONS).
  - lock-order / reaper-double-free / cross-CPU-bounds: clean.

Out of scope (deferred decision, signed off): the CRITICAL the workflow
surfaced — task slots are not a globally-arbitrated namespace (per-CPU-local
free-slot search), so concurrent cross-CPU SYS_SPAWN can collide on
(slot, generation). This is PRE-EXISTING (the bare-u32 slot already collided,
mis-routing wakes; migrate_task_between already drops a task on accept
failure); Phase B inherits the "slots are global" assumption that TASK_CPU_MAP
and migration already make, without worsening the danger, and today's
single-shell spawner never triggers it. A Convention-9 deferral at the
allocation site (src/scheduler/mod.rs create_isr_task) records the assumption
and the trigger (a second concurrent cross-CPU spawner). The fix is a focused
global task-slot allocator landing as the next change. STATUS.md known-issue
to follow (separate).

Verification:
  - make check-all: x86_64 + aarch64 + riscv64 build cleanly.
  - cargo test --lib: 900 passed, 0 failed (8 new).
  - check-assumptions / check-deferrals / check-unsafe-coverage /
    check-boot-panics: all pass (0 new; the deferral carries its trigger,
    baselines carry only line-drift). clippy: no new warnings.
  - make run-quiet: boots to CambiOS Shell (exit 0, 0 scheduler anomalies) —
    fde-mount's boot exit now also reaps its slot + bumps generation cleanly.
  - The full spawn->exit->wait round-trip is best confirmed interactively
    (play <game> + Ctrl+Q -> "exited with code N"); run-quiet cannot inject
    input.

Staged files:
  docs/ASSUMPTIONS.md
  src/lib.rs
  src/reaper.rs
  src/scheduler/SCHEDULER.md
  src/scheduler/mod.rs
  src/scheduler/task.rs
  src/syscalls/dispatcher.rs
  tools/check-assumptions-baseline.txt
  tools/check-deferrals-baseline.txt
  user/libsys/src/lib.rs
  user/shell/src/main.rs
  user/terminal-window/src/main.rs

+++ user/terminal-window/src/main.rs
@@ -301,7 +301,8 @@ fn cmd_play(t: &mut Terminal<GuiBackend>, args: &[u8]) {
         return;
     }
 
-    let _exit = sys::wait_task(tid as u32);
+    // Full (slot, generation) handle — must not truncate to the slot (ADR-034).
+    let _exit = sys::wait_task(tid as u64);
 
     // Game's done — re-attach our window. If reopen fails the GUI is
     // dead and the terminal can't draw; clean exit beats a zombie.
